Avoid static state
Motivation:

When statically compiled, the same netty-jni-util may be used by
multiple separate JNI libraries. In this situation there is only one
copy of static variables, so it is only safe to use static variables in
cases where they can safely be reused. That's not the case for the
current staticPackagePrefix.

Modifications:

Force the caller to manage the static state, which can be unique per
usage.

Result:

Static compilation is available again for JNI libraries that supported
static compilation before depending on netty-jni-util. Fixes #5.