Avoid use of strdup, strndup (#8)
Motivation::

On old glibc, strdup and strndup are aliased to __strdup and __strndup,
which musl does not implement. This causes netty to unnecessarily fail
on musl. strndup is not the right tool here anyways, since we know the
exact size of the string to be duplicated, not the maximum size.

Modification:

* Don't call strdup, strndup
* Don't free packagePrefix after load_function

Result:

Better compat